step4_wscn/requirements.txt
readability-lxml
jieba